ICU-5966 Fix some compiler warnings.
X-SVN-Rev: 23540

+++ icu4c/source/tools/makeconv/genmbcs.c
@@ -1,7 +1,7 @@
 /*
 *******************************************************************************
 *
-*   Copyright (C) 2000-2007, International Business Machines
+*   Copyright (C) 2000-2008, International Business Machines
 *   Corporation and others.  All Rights Reserved.
 *
 *******************************************************************************
@@ -1431,8 +1431,10 @@ MBCSWrite(NewConverter *cnvData, const UConverterStaticData *staticData,
         stage2Length-=stage2Start;
         if(VERBOSE) {
             printf("+ omitting %lu out of %lu stage2 entries and %lu fromUBytes\n",
-                   stage2Start, mbcsData->stage2Top, mbcsData->stage3Top);
-            printf("+ total size savings: %lu bytes\n", stage2Start*4+mbcsData->stage3Top);
+                    (unsigned long)stage2Start,
+                    (unsigned long)mbcsData->stage2Top,
+                    (unsigned long)mbcsData->stage3Top);
+            printf("+ total size savings: %lu bytes\n", (unsigned long)stage2Start*4+mbcsData->stage3Top);
         }
     } else {
         stage2Start=0;
